fix: rename and move PreloadResources
